+++ CHANGELOG.md
@@ -10,10 +10,14 @@ All notable changes to this project will be documented in this file.
 
 - Added reusable Polars expression helper functions in `polars_utils/expressions.py` for `is_care_home()`, `is_not_care_home()`, and `is_dormant()`.
 
+- Converted filter for job group outliers to polars.
+
 - In line Polars conversions into the Impute job for the PySpark functions `combine_care_home_and_non_res_values_into_single_column` and `convert_care_home_ratios_to_posts`.
 
 ### Changed
 - Renamed replace_care_navigator_with_care_coordinator to remap_mainjrid_codes for clarity and optimised implementation to use dictionary-based replacements targeting the main_job_role_clean column and updated corresponding test data names. Added Technician Job Role('22') in the dictionary to replace it with Other non-care-providing job roles ('27').
+
+- Changed apply_categorical_labels so it can reverse labels by swapping the k:v in the label dict.
 
 ### Fixed